+++ source/js/components/Dashboard/Sidebar/Filter/FilterTitle.jsx
@@ -17,9 +17,10 @@ import { FormattedMessage } from 'react-intl';
     showFilterMenu: state.menu.get('showFilterMenu'),
     filtersActive,
   };
-})
-class FilterButton extends Component {
+}, null, null, { pure: false })
+class FilterTitle extends Component {
   static propTypes = {
+    items: PropTypes.object,
     showFilterMenu: PropTypes.bool,
     filtersActive: PropTypes.bool,
 
@@ -39,28 +40,44 @@ class FilterButton extends Component {
   }
 
   render() {
-    const { showFilterMenu, filtersActive } = this.props;
+    const { items, showFilterMenu, filtersActive } = this.props;
 
-    const ItemClass = classNames({
+    const ButtonClass = classNames({
       'FilterButton': true,
       'FilterButton--open': showFilterMenu,
       'FilterButton--filtersActive': filtersActive,
     });
 
     return (
-      <button
-        className={ ItemClass }
-        onClick={ this.toggle }
-      >
-        <FormattedMessage
-          id='filter.filterButtonTitle'
-          description='Title of the Filter Menu Toggle-Button'
-          defaultMessage='Liste filtern'
-        />
-      </button>
+      <header className='FilterTitle'>
+        <h2>
+          <FormattedMessage
+            id='filter.itemCount'
+            description='Title of the Filter Menu'
+            defaultMessage='{itemCount, plural,
+                =0 {keine Objekte}
+                one {1 Objekte}
+                other {{itemCount} Objekte}
+            }'
+            values={ {
+              itemCount: items ? items.size : 0,
+            } }
+          />
+        </h2>
+        <button
+          className={ ButtonClass }
+          onClick={ this.toggle }
+        >
+          <FormattedMessage
+            id='filter.filterButtonTitle'
+            description='Title of the Filter Menu Toggle-Button'
+            defaultMessage='Liste filtern'
+          />
+        </button>
+      </header>
     );
   }
 
 }
 
-export default FilterButton;
+export default FilterTitle;
