+++ source/scss/base/_fonts.scss
@@ -1,3 +1,7 @@
 @import url('https://fonts.googleapis.com/css?family=Space+Mono:400,700');
 
 $main-font: "Space Mono", "Times New Roman", serif;
+
+h2 {
+  font-size: 1.5em;
+}
