+++ source/scss/components/_filters.scss
@@ -1,3 +1,31 @@
+.FilterTitle {
+  display: flex;
+  padding: 1rem 0.5rem;
+
+  h2 {
+    flex: 1;
+    margin: 0;
+    text-shadow:
+      2px 2px 0 white,
+      0 2px 0 white,
+      2px 0 0 white,
+      -2px -2px 0 white,
+      0 -2px 0 white,
+      -2px 0 0 white;
+  }
+
+  button {
+    background-color: $primary-color;
+    color: white;
+    transition: background-color .2s ease;
+    padding: 0 0.5em;
+
+    &:hover {
+      background-color: darken($primary-color, 10%);
+    }
+  }
+}
+
 .FilterSection {
   position:fixed;
   top:0;
@@ -15,17 +43,22 @@
 
   &.FilterSection--active {
     transform: translateY(0);
+
+    @include md {
+      transform: translateX(0);
+    }
   }
 
   @include md {
-    transform: translateY(0);
-    position:relative;
-    padding:0;
+    transform: translateX(-110%);
     z-index: $z-index-menu - 10;
-
-    h1, h2 {
-      display: none;
-    }
+    margin: 0 0.5rem;
+    left: -0.5rem;
+    right: auto;
+    top: $header-height;
+    width: calc(33% + 0.5rem);
+    box-shadow:.1em .1em .3em rgba(0,0,0,.15);
+    z-index: $z-index-menu - 5;
   }
 
   h1 {
@@ -42,10 +75,11 @@
       background-color: $primary-color-contrast;
       font-weight:bold;
       text-transform: uppercase;
-    }
 
-    @include md {
-      display: none;
+      @include md {
+        margin-top: 1rem;
+        font-size: 1.4em;
+      }
     }
   }
 }
@@ -55,18 +89,9 @@
   flex-wrap: wrap;
   justify-content: start;
 
-  @include md {
-    background-color: rgb(100,100,100);
-  }
-
   .CategoryFilter-Item {
     width: 100%;
     margin-bottom: .5em;
-
-    @include md {
-      margin-bottom:0;
-      width: 16.666%;
-    }
 
     button{
       position:relative;
@@ -79,23 +104,7 @@
       text-align:left;
 
       @include md {
-        text-align: center;
-        padding-left:0;
-
-        &:after {
-          display: none;
-        }
-
-        &:before {
-          content:'';
-          position:absolute;
-          top:0;
-          left:0;
-          right:0;
-          bottom:0;
-          background-color:rgba(0,0,0,.6);
-          transition: opacity .2s;
-        }
+        padding: 0.5em 0 0.5em 0.5em;
       }
     }
 
@@ -144,12 +153,7 @@
     transition: background-color .2s ease;
 
     @include md {
-      width: 33%;
-      border-right: 1px solid rgba(255,255,255,.2);
-      flex:1;
-      margin-bottom:0;
-      text-align:center;
-      padding-left:0;
+      padding: 0.5em 0 0.5em 0.5em;
     }
 
     &:last-child {
@@ -174,6 +178,13 @@
     background-repeat:no-repeat;
     background-position:center;
     background-size: contain;
+
+    @include md {
+      right: 0.5em;
+      width: 1.5em;
+      height: 1.5em;
+      bottom: 0.6em;
+    }
   }
 }
