+++ source/translations/locales/whitelist_de.json
@@ -7,15 +7,14 @@
   "filter.incomplete",
   "filter.missingImages",
   "filter.missingDescription",
+  "map.switchMunicipality",
   "uploadPhoto",
   "cancel",
-  "cityinfo.openLink",
   "item.photoLinkTitle",
   "item.readArticle",
   "item.createArticle",
   "item.missingTitle",
   "item.missingDescription",
-  "map.switchMunicipality",
   "search.viennaWarning",
   "search.placeholder",
   "header.logoAlt",
